Node_JS_tutorial/old_qlts_project/requirements2.txt
asgiref==3.6.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.18
django-braces==1.15.0
django-filter==23.1
django-import-export==3.2.0
django-ipware==5.0.0
django-model-tracker==2.0.3
django-model-utils==1.4.0
django-multiselectfield==0.1.12
django-netfields==1.3.1
django-oauth-toolkit==2.2.0
django-rest-framework-social-oauth2==1.1.0
django-tracking-fields==1.3.7
django-watchman==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==3.4
importlib-metadata==6.4.1
install==1.3.5
jwcrypto==1.4.2
Markdown==3.4.3
MarkupPy==1.14
netaddr==0.8.0
oauth-provider==1.1
oauth2-provider==0.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
psutil==5.9.5
psycopg2-binary==2.9.6
py-cpuinfo==9.0.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
resources==0.0.1
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.4
tablib==3.4.0
typing_extensions==4.5.0
urllib3==1.26.15
whitenoise==6.4.0
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
